// backup/auto-backup-2025-09-10-1123/src/types/models.ts
export interface Favorite {
  id: string;
  userId: string;
  
  // Favorite Target
  targetType: FavoriteTargetType;
  targetId: string;
  targetName: string;
  targetImageUrl: string;
  
  // Context
  shopId: string;
  shopName: string;
  
  // Metadata
  notes: string;
  tags: string[];
  
  // Timestamps
  createdAt: number;
}

export enum FavoriteTargetType {
  SHOP = 'SHOP',
  PRODUCT = 'PRODUCT',
  SERVICE = 'SERVICE'
}